+++ lxmls/multimodal/gemma3/model.py
@@ -11,20 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Inference-only Gemma model implementation."""
 
 import gc
 import json
 import os
-from typing import Any, List, Mapping, Optional, Sequence, Tuple, Union
+from typing import List, Mapping, Optional, Tuple, Union
 
 import torch
 import torch.nn.functional as F
-from PIL import Image
 from torch import nn
 
-from lxmls.multimodal.gemma3 import preprocessor, siglip_vision, tokenizer
+from lxmls.multimodal.gemma3 import siglip_vision
 from lxmls.multimodal.gemma3.config import AttentionType, GemmaConfig
+from lxmls.multimodal.gemma3.processor import Tokenizer
 
 
 class Sampler(nn.Module):
@@ -375,213 +374,6 @@ class GemmaTextModel(nn.Module):
         return hidden_states
 
 
-class GemmaForCausalLM(nn.Module):
-    def __init__(self, config: GemmaConfig):
-        super().__init__()
-        self.config = config
-        assert config.hidden_size % config.num_attention_heads == 0
-
-        max_seq_len = config.max_position_embeddings
-        head_dim = config.head_dim
-        vocab_size = config.vocab_size
-
-        self.tokenizer = tokenizer.Tokenizer(config.tokenizer)
-        self.embedder = Embedding(vocab_size, config.hidden_size, config.quant)
-        self.model = GemmaTextModel(config)
-        self.sampler = Sampler(vocab_size, config)
-
-        # Pre-compute rotary embedding table.
-        if config.rope_wave_length is None:
-            raise ValueError("rope_wave_length must be provided for Gemma3.")
-
-        rope_lengths = config.rope_wave_length
-        defaults = {
-            AttentionType.LOCAL_SLIDING: 10_000,
-            AttentionType.GLOBAL: 10_000,
-        }
-
-        for attn_type, name in [
-            (AttentionType.LOCAL_SLIDING, "local_freqs_cis"),
-            (AttentionType.GLOBAL, "global_freqs_cis"),
-        ]:
-            theta = rope_lengths.get(attn_type, defaults[attn_type])
-            self._register_freqs_cis(name, head_dim, max_seq_len, theta=theta)
-
-    def _register_freqs_cis(self, name: str, head_dim: int, max_seq_len: int, theta: int = 10_000):
-        self.register_buffer(name, precompute_freqs_cis(head_dim, max_seq_len * 2, theta=theta))
-
-    @torch.no_grad()
-    def forward(
-        self,
-        input_token_ids: torch.Tensor,
-        input_positions: torch.Tensor,
-        kv_write_indices: torch.Tensor,
-        kv_caches: List[Tuple[torch.Tensor, torch.Tensor]],
-        mask: torch.Tensor,
-        output_positions: torch.Tensor,
-        temperatures: Union[torch.Tensor, None],
-        top_ps: torch.Tensor,
-        top_ks: torch.Tensor,
-        local_mask: torch.Tensor | None = None,
-        **kwargs,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        freqs_cis = {}
-
-        freqs_cis[AttentionType.LOCAL_SLIDING] = self.local_freqs_cis.index_select(0, input_positions)
-        freqs_cis[AttentionType.GLOBAL] = self.global_freqs_cis.index_select(0, input_positions)
-
-        kv_write_indices = input_positions
-
-        # [batch_size, input_len, hidden_size]
-        hidden_states = self.embedder(input_token_ids)
-        # Gemma normalizes the embedding by sqrt(hidden_size).
-        # Gemma2 downcasts the below to float16, causing sqrt(3072)=55.4256 to become 55.5
-        # See https://github.com/huggingface/transformers/pull/29402
-        normalizer = torch.tensor(self.config.hidden_size**0.5, dtype=hidden_states.dtype, device=hidden_states.device)
-        hidden_states = hidden_states * normalizer
-
-        hidden_states = self.model(
-            hidden_states=hidden_states,
-            freqs_cis=freqs_cis,
-            kv_write_indices=kv_write_indices,
-            kv_caches=kv_caches,
-            mask=mask,
-            local_mask=local_mask,
-        )
-        embedder_weight = self.embedder.weight
-        if self.config.quant:
-            embedder_weight = embedder_weight * self.embedder.weight_scaler.unsqueeze(-1)
-        next_tokens, logits = self.sampler(
-            embedding=embedder_weight,
-            hidden_states=hidden_states,
-            output_positions=output_positions,
-            temperatures=temperatures,
-            top_ps=top_ps,
-            top_ks=top_ks,
-        )
-        return next_tokens, logits
-
-    def generate(
-        self,
-        prompts: Union[str, Sequence[str]],
-        device: Any,
-        output_len: int = 100,
-        temperature: Union[float, None] = 1.0,
-        top_p: float = 0.95,
-        top_k: int = 64,
-    ) -> Union[str, Sequence[str]]:
-        """Generates responses for given prompts using Gemma model."""
-        # If a single prompt is provided, treat it as a batch of 1.
-        is_str_prompt = isinstance(prompts, str)
-        if is_str_prompt:
-            prompts = [prompts]
-
-        batch_size = len(prompts)
-        prompt_tokens = [self.tokenizer.encode(prompt) for prompt in prompts]
-        min_prompt_len = min(len(p) for p in prompt_tokens)
-        max_prompt_len = max(len(p) for p in prompt_tokens)
-        max_seq_len = max_prompt_len + output_len
-        assert max_seq_len <= self.config.max_position_embeddings
-
-        # build KV caches
-        kv_caches = []
-        for _ in range(self.config.num_hidden_layers):
-            size = (batch_size, max_seq_len, self.config.num_key_value_heads, self.config.head_dim)
-            dtype = self.config.get_dtype()
-            k_cache = torch.zeros(size=size, dtype=dtype, device=device)
-            v_cache = torch.zeros(size=size, dtype=dtype, device=device)
-            kv_caches.append((k_cache, v_cache))
-
-        # prepare inputs
-        token_ids_tensor = torch.full((batch_size, max_seq_len), self.tokenizer.pad_id, dtype=torch.int64)
-        input_token_ids_tensor = torch.full((batch_size, min_prompt_len), self.tokenizer.pad_id, dtype=torch.int64)
-        for i, p in enumerate(prompt_tokens):
-            token_ids_tensor[i, : len(p)] = torch.tensor(p)
-            input_token_ids_tensor[i, :min_prompt_len] = torch.tensor(p[:min_prompt_len])
-        token_ids_tensor = token_ids_tensor.to(device)
-        input_token_ids_tensor = input_token_ids_tensor.to(device)
-        prompt_mask_tensor = token_ids_tensor != self.tokenizer.pad_id
-        input_positions_tensor = torch.arange(0, min_prompt_len, dtype=torch.int64).to(device)
-        mask_tensor = torch.full((1, 1, max_seq_len, max_seq_len), -2.3819763e38).to(torch.float)
-        mask_tensor = torch.triu(mask_tensor, diagonal=1).to(device)
-        local_mask_tensor = (
-            mask_tensor
-            + torch.tril(
-                torch.full((1, 1, max_seq_len, max_seq_len), -2.3819763e38, device=device),
-                diagonal=-self.config.sliding_window_size,
-            )
-            if self.config.sliding_window_size
-            else None
-        )
-        curr_mask_tensor = mask_tensor.index_select(dim=2, index=input_positions_tensor)
-        curr_local_mask_tensor = (
-            local_mask_tensor.index_select(2, input_positions_tensor) if local_mask_tensor is not None else None
-        )
-        output_positions_tensor = torch.LongTensor([min_prompt_len - 1]).to(device)
-        temperatures_tensor = None if not temperature else torch.FloatTensor([temperature] * batch_size).to(device)
-        top_ps_tensor = torch.FloatTensor([top_p] * batch_size).to(device)
-        top_ks_tensor = torch.LongTensor([top_k] * batch_size).to(device)
-        output_index = torch.tensor(min_prompt_len, dtype=torch.int64).to(device)
-
-        # Prefill up to min_prompt_len tokens, then treat other prefill as
-        # decode and ignore output.
-        for i in range(max_seq_len - min_prompt_len):
-            next_token_ids, _ = self(
-                input_token_ids=input_token_ids_tensor,
-                input_positions=input_positions_tensor,
-                kv_write_indices=None,
-                kv_caches=kv_caches,
-                mask=curr_mask_tensor,
-                output_positions=output_positions_tensor,
-                temperatures=temperatures_tensor,
-                top_ps=top_ps_tensor,
-                top_ks=top_ks_tensor,
-                local_mask=curr_local_mask_tensor,
-            )
-            curr_prompt_mask = prompt_mask_tensor.index_select(dim=1, index=output_index).squeeze(dim=1)
-            curr_token_ids = token_ids_tensor.index_select(dim=1, index=output_index).squeeze(dim=1)
-            output_token_ids = torch.where(curr_prompt_mask, curr_token_ids, next_token_ids).unsqueeze(dim=1)
-            token_ids_tensor.index_copy_(dim=1, index=output_index, source=output_token_ids)
-
-            input_token_ids_tensor = output_token_ids
-            input_positions_tensor = output_index.unsqueeze(dim=-1)
-            curr_mask_tensor = mask_tensor.index_select(2, input_positions_tensor)
-            curr_local_mask_tensor = (
-                local_mask_tensor.index_select(2, input_positions_tensor) if local_mask_tensor is not None else None
-            )
-
-            output_positions_tensor = torch.tensor(0, dtype=torch.int64).to(device)
-            output_index = output_index + 1
-
-        # Detokenization.
-        token_ids = token_ids_tensor.tolist()
-        results = []
-        for i, tokens in enumerate(token_ids):
-            trimmed_output = tokens[len(prompt_tokens[i]) : len(prompt_tokens[i]) + output_len]
-            if self.tokenizer.eos_id in trimmed_output:
-                eos_index = trimmed_output.index(self.tokenizer.eos_id)
-                trimmed_output = trimmed_output[:eos_index]
-            results.append(self.tokenizer.decode(trimmed_output))
-
-        # If a string was provided as input, return a string as output.
-        return results[0] if is_str_prompt else results
-
-    def load_weights(self, model_path: str):
-        if os.path.isfile(model_path):
-            self.load_state_dict(torch.load(model_path, mmap=True, weights_only=True)["model_state_dict"], strict=False)
-        else:
-            index_path = os.path.join(model_path, "pytorch_model.bin.index.json")
-            with open(index_path, "r", encoding="utf-8") as f:
-                index = json.load(f)
-            shard_files = list(set(index["weight_map"].values()))
-            for shard_file in shard_files:
-                shard_path = os.path.join(model_path, shard_file)
-                state_dict = torch.load(shard_path, map_location="cpu", weights_only=True)
-                self.load_state_dict(state_dict, strict=False)
-                del state_dict  # Save memory.
-                gc.collect()
-
-
 class Gemma3ForMultimodalLM(nn.Module):
     def __init__(self, config: GemmaConfig):
         super().__init__()
@@ -590,7 +382,7 @@ class Gemma3ForMultimodalLM(nn.Module):
         max_seq_len = config.max_position_embeddings
         head_dim = config.head_dim
         vocab_size = config.vocab_size
-        self.tokenizer = tokenizer.Tokenizer(config.tokenizer)
+        self.tokenizer = Tokenizer(config.tokenizer)
         self.text_token_embedder = Embedding(vocab_size, config.hidden_size, config.quant)
         self.model = GemmaTextModel(config)
         self.sampler = Sampler(vocab_size, config)
@@ -760,120 +552,6 @@ class Gemma3ForMultimodalLM(nn.Module):
             ),
         )
         return attention_mask, local_mask
-
-    def generate(
-        self,
-        prompts: Sequence[Sequence[Union[str, Image.Image]]],
-        device: Any,
-        output_len: int = 100,
-        temperature: Union[float, None] = 1.0,
-        top_p: float = 0.95,
-        top_k: int = 64,
-    ) -> Sequence[str]:
-        processing_result = preprocessor.tokenize_raw_input(self.tokenizer, prompts, self.config, output_len, device)
-        batch_size = processing_result.batch_size
-        finalised_token_ids = processing_result.finalised_token_ids
-        image_batch = processing_result.image_batch
-        min_prompt_len = processing_result.min_prompt_len
-        _max_prompt_len = processing_result.max_prompt_len
-        total_seq_len = processing_result.max_seq_len
-        image_presence_mask = processing_result.image_presence_mask
-
-        # Create attention mask.
-        min_dtype = torch.finfo(self.dtype).min
-        n_inf = torch.tensor(min_dtype, dtype=torch.float32, device=device)
-        assert self.config.sliding_window_size is not None
-        boolean_mask, local_boolean_mask = self.create_attention_mask(finalised_token_ids, total_seq_len)
-        mask_tensor = torch.where(boolean_mask, 0, n_inf).contiguous()
-        local_mask_tensor = torch.where(local_boolean_mask, 0, n_inf).contiguous()
-
-        kv_caches = []
-        for _ in range(self.config.num_hidden_layers):
-            size = (batch_size, total_seq_len, self.config.num_key_value_heads, self.config.head_dim)
-            dtype = self.config.get_dtype()
-            k_cache = torch.zeros(size=size, dtype=dtype, device=device)
-            v_cache = torch.zeros(size=size, dtype=dtype, device=device)
-            kv_caches.append((k_cache, v_cache))
-
-        input_token_ids_tensor = torch.full(
-            (batch_size, min_prompt_len), self.tokenizer.pad_id, dtype=torch.int64, device=device
-        )
-        token_ids_tensor = finalised_token_ids.to(device)
-        for i in range(batch_size):
-            p = finalised_token_ids[i]
-            input_token_ids_tensor[i, :min_prompt_len] = p[:min_prompt_len]
-
-        input_positions_tensor = torch.arange(0, min_prompt_len, dtype=torch.int64, device=device)
-        prompt_mask_tensor = token_ids_tensor != self.tokenizer.pad_id
-        curr_mask_tensor = mask_tensor.index_select(dim=2, index=input_positions_tensor)
-        curr_local_mask_tensor = local_mask_tensor.index_select(dim=2, index=input_positions_tensor)
-        # The first iteration produces a sequence of hidden states of shape: [B, min_prompt_len, hidden_size]
-        # To predict the next token, we need to look at the hidden state corresponding to the last token: [min_prompt_len - 1]
-        output_positions_tensor = torch.LongTensor([min_prompt_len - 1]).to(device)
-        temperatures_tensor = None if not temperature else torch.FloatTensor([temperature] * batch_size).to(device)
-        top_ps_tensor = torch.FloatTensor([top_p] * batch_size).to(device)
-        top_ks_tensor = torch.LongTensor([top_k] * batch_size).to(device)
-        output_index = torch.tensor(min_prompt_len, dtype=torch.int64, device=device)
-
-        # Prefill up to min_prompt_len tokens, then treat other prefill as
-        # decode and ignore output.
-        for i in range(total_seq_len - min_prompt_len):
-            next_token_ids, _ = self(
-                input_token_ids=input_token_ids_tensor,
-                image_patches=image_batch,  # NOTE: Always `None` after the first iteration
-                image_presence_mask=image_presence_mask,
-                input_positions=input_positions_tensor,
-                kv_caches=kv_caches,
-                mask=curr_mask_tensor,
-                output_positions=output_positions_tensor,
-                temperatures=temperatures_tensor,
-                top_ps=top_ps_tensor,
-                top_ks=top_ks_tensor,
-                local_mask=curr_local_mask_tensor,
-            )
-            # Check if the current generation step corresponds to a position that was part of the original prompt
-            curr_prompt_mask = prompt_mask_tensor.index_select(dim=1, index=output_index).squeeze(dim=1)
-            # Retrieve the token from the original prompt at the current position
-            curr_token_ids = token_ids_tensor.index_select(dim=1, index=output_index).squeeze(dim=1)
-            # Decide which token to use
-            # if `curr_prompt_mask` is True (i.e. we are still processing the prompt) use prompt token curr_token_ids
-            # else use the token generated by the model: next_token_ids
-            output_token_ids = torch.where(curr_prompt_mask, curr_token_ids, next_token_ids).unsqueeze(dim=1)
-            # Update the 'final' copy of tokens
-            token_ids_tensor.index_copy_(dim=1, index=output_index, source=output_token_ids)
-
-            # Token just generated (or prefilled) is not part of the input
-            input_token_ids_tensor = output_token_ids
-            # The input position for the next iteration is set to the current output_index
-            input_positions_tensor = output_index.unsqueeze(dim=-1)
-            # The attention masks are updated to reflect the new seq_len for the next iteration
-            curr_mask_tensor = mask_tensor.index_select(2, input_positions_tensor)
-            curr_local_mask_tensor = (
-                local_mask_tensor.index_select(2, input_positions_tensor) if local_mask_tensor is not None else None
-            )
-            # After the first iteration, where we pass in min_prompt_len tokens, we now generate tokens one by one
-            # The input to the model now becomes a single token
-            # The model processes this token and produces a hidden state sequence of shape: [B, 1, hidden_size]
-            # So, now to get the logits for the next token, we need to use the first (and only) hidden state available
-            output_positions_tensor = torch.tensor(0, dtype=torch.int64, device=device)
-            # duh
-            output_index = output_index + 1
-
-            # The image information is already integrated into the text embeddings in the first iteration
-            image_batch = None
-            image_presence_mask = None
-
-        # Detokenization.
-        token_ids = token_ids_tensor.tolist()
-        results = []
-        for i, tokens in enumerate(token_ids):
-            output = tokens
-            if self.tokenizer.eos_id in output:
-                eos_index = output.index(self.tokenizer.eos_id)
-                output = output[:eos_index]
-            results.append(self.tokenizer.decode(output))
-
-        return results
 
     def load_weights(self, model_path: str):
         if os.path.isfile(model_path):
