--- lxmls/multimodal/gemma3/preprocessor.py
@@ -1,266 +1,0 @@
-# Copyright 2024 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from collections.abc import Sequence
-from dataclasses import dataclass
-from typing import Any, Optional, Union
-
-import numpy as np
-import torch
-from PIL import Image
-
-import lxmls.multimodal.gemma3.config as gemma_config
-from lxmls.multimodal.gemma3.siglip_vision import preprocess_images_for_siglip_vision
-from lxmls.multimodal.gemma3.tokenizer import Tokenizer
-
-CROPPED_IMAGE_PREFIX = "here is the original image"
-CROPPED_IMAGE_FILLER = "and here are some crops to help you see better"
-
-
-@dataclass
-class TokenisationOutput:
-    finalised_token_ids: torch.Tensor
-    image_batch: Optional[torch.Tensor]
-    batch_size: int
-    min_prompt_len: int
-    max_prompt_len: int
-    max_seq_len: int
-    image_presence_mask: Optional[torch.Tensor]
-
-
-def pan_and_scan(img: Image.Image, *, min_crop_size: int = 256, max_num_crops: int = 4) -> Sequence[Image.Image]:
-    """Pan and scan an image for open source.
-
-    If the image is landscape, the crops are made horizontally and if the image is
-    portrait, the crops are made vertically. The longer side of the image is split
-    into [2 - max_num_crops] crops.
-
-    Args:
-        img: PIL Image object.
-        min_crop_size: The minimum size of each crop.
-        max_num_crops: The maximum desired number of crops to be generated.
-
-    Returns:
-        List of cropped PIL Image objects and a list of crop positions.
-    """
-    w, h = img.size
-
-    # Square or landscape image.
-    if w >= h:
-        if w / h < 1.5:
-            # return [img], [(0, 0, h, w)]
-            return [img]
-
-        # Select ideal number of crops close to the image aspect ratio and such that
-        # crop_size > min_crop_size.
-        num_crops_w = int(np.floor(w / h + 0.5))  # Half round up rounding.
-        num_crops_w = min(int(np.floor(w / min_crop_size)), num_crops_w)
-
-        # Make sure the number of crops is in range [2, max_num_crops].
-        num_crops_w = max(2, num_crops_w)
-        num_crops_w = min(max_num_crops, num_crops_w)
-        num_crops_h = 1
-
-    # Portrait image.
-    else:
-        if h / w < 1.5:
-            # return [img], [(0, 0, h, w)]
-            return [img]
-
-        num_crops_h = int(np.floor(h / w + 0.5))
-        num_crops_h = min(int(np.floor(h / min_crop_size)), num_crops_h)
-        num_crops_h = max(2, num_crops_h)
-        num_crops_h = min(max_num_crops, num_crops_h)
-        num_crops_w = 1
-
-    crop_size_w = int(np.ceil(w / num_crops_w))
-    crop_size_h = int(np.ceil(h / num_crops_h))
-
-    # Don't apply pan and scan if crop size is too small.
-    if min(crop_size_w, crop_size_h) < min_crop_size:
-        # return [img], [(0, 0, h, w)]
-        return [img]
-
-    crop_positions_w = [crop_size_w * i for i in range(num_crops_w)]
-    crop_positions_h = [crop_size_h * i for i in range(num_crops_h)]
-
-    # Generate crops.
-    crops = []
-    crop_positions = []
-    for pos_h in crop_positions_h:
-        for pos_w in crop_positions_w:
-            crops.append(img.crop((pos_w, pos_h, pos_w + crop_size_w, pos_h + crop_size_h)))
-            crop_positions.append((pos_h, pos_w, pos_h + crop_size_h, pos_w + crop_size_w))
-
-    # return crops, crop_positions
-    return crops
-
-
-def input_preprocessor(
-    raw_user_prompt: Sequence[Union[Image.Image, str]],
-) -> Sequence[Union[torch.Tensor, str]]:
-    """Preprocessor for Gemma3 input
-
-    Args:
-      raw_user_prompt: A list of images or strings, as provided by the user.
-
-    Returns:
-      A list of preprocessed images or strings.
-    """
-    preprocessed_input: list[Union[torch.Tensor, str]] = []
-    for element in raw_user_prompt:
-        if isinstance(element, Image.Image):
-            cropped_images = pan_and_scan(element)
-            preprocessed_images_cropped = preprocess_images_for_siglip_vision(cropped_images)
-            preprocessed_images_uncropped = preprocess_images_for_siglip_vision([element])
-            if len(preprocessed_images_cropped) == 1:
-                preprocessed_input.append(preprocessed_images_uncropped[0])
-            elif len(preprocessed_images_cropped) > 1:
-                preprocessed_input.append(CROPPED_IMAGE_PREFIX)
-                preprocessed_input.append(preprocessed_images_uncropped[0])
-                preprocessed_input.append(CROPPED_IMAGE_FILLER)
-                preprocessed_input.extend(preprocessed_images_cropped)
-            else:
-                raise ValueError("No images found in the input.")
-        else:
-            preprocessed_input.append(element)
-
-    return preprocessed_input
-
-
-def batch_input_preprocessor(raw_input: Sequence[Sequence[Union[Image.Image, str]]]):
-    """Preprocessor for Gemma3 batch input"""
-    preprocessed_input: list[Sequence[Union[torch.Tensor, str]]] = []
-    for element in raw_input:
-        preprocessed_input.append(input_preprocessor(element))
-    return preprocessed_input
-
-
-def tokenize_raw_input(
-    tokenizer: Tokenizer,
-    raw_input: Sequence[Sequence[Union[str, Image.Image]]],
-    config: gemma_config.GemmaConfig,
-    output_len: int,
-    device: Any,
-) -> TokenisationOutput:
-    """
-    Converts a preprocessed batch of interleaved text and image inputs into
-    token IDs and an image batch suitable for gemma3 model
-
-    Args:
-        preprocessed_batch: List of lists containing strings and torch.Tensor images.
-        image_token_id: Token ID to represent image placeholders.
-        max_image_tokens: Number of tokens reserved for each image.
-        image_size: Expected size of images (C, H, W).
-
-    Returns:
-        user_input_token_ids: Batch of token IDs with shape (B, L), where L is the max sequence length.
-        image_batch: Batch of images with shape (B, N, C, H, W), where N is the max number of images.
-    """
-    vis_cfg = config.vision_config
-    assert vis_cfg is not None
-
-    preprocessed_batch = batch_input_preprocessor(raw_input)
-
-    # Initialize lists to store token IDs and image tensors
-    all_token_ids, all_images, prompt_lengths = [], [], []
-    min_prompt_len, max_prompt_len, max_num_images = float("inf"), 0, 0
-
-    # Iterate over each user prompt in the batch
-    for prompt in preprocessed_batch:
-        token_ids, images = [], []
-        token_ids.append(tokenizer.bos_id)
-        # Process each element in the prompt
-        for element in prompt:
-            if isinstance(element, str):
-                # Tokenize text and add to token_ids
-                tokens = tokenizer.encode(element, bos=False, eos=False)
-                token_ids.extend(tokens)
-            elif isinstance(element, torch.Tensor):
-                # Prepend (dual endline + [BEGIN OF IMAGE])
-                token_ids.extend(tokenizer.encode("\n\n", bos=False, eos=False))
-                token_ids.append(tokenizer.boi_id)
-                # Add image placeholder tokens
-                token_ids.extend([tokenizer.image_token_placeholder_id] * vis_cfg.encoding_sequence_length)
-                # Append ([END OF IMAGE] + dual endline)
-                token_ids.append(tokenizer.eoi_id)
-                token_ids.extend(tokenizer.encode("\n\n", bos=False, eos=False))
-                # Store the image tensor
-                images.append(element)
-            else:
-                raise ValueError("Unsupported type in prompt. Expected str or torch.Tensor.")
-
-        # Update values
-        curr_prompt_len = len(token_ids)
-        prompt_lengths.append(curr_prompt_len)
-        max_prompt_len = max(max_prompt_len, curr_prompt_len)
-        min_prompt_len = min(min_prompt_len, curr_prompt_len)
-        max_num_images = max(max_num_images, len(images))
-
-        all_token_ids.append(token_ids)
-        all_images.append(images)
-
-    max_seq_len = max_prompt_len + output_len
-
-    # Pad token IDs to the maximum sequence length
-    finalised_token_ids = []
-    for token_ids in all_token_ids:
-        pad_length = max_seq_len - len(token_ids)
-        padded_token_ids = token_ids + [tokenizer.pad_id] * pad_length
-        finalised_token_ids.append(padded_token_ids)
-
-    # Pad images to the maximum number of images in the batch
-    image_batch = []
-    image_presence_mask = []
-    for images in all_images:
-        # Check if all images within the current sublist have the same shape
-        if images:  # Check if the sublist is not empty
-            first_shape = images[0].shape
-            for img in images:
-                assert img.shape == first_shape, "Images within a sublist must have the same shape."
-        pad_length = max_num_images - len(images)
-        padded_images = images.copy()  # create a copy so the original data is not altered.
-        presence_mask = [True] * len(images)
-
-        if pad_length > 0:
-            # Create a list of zero tensors for padding
-            padding = [
-                torch.zeros((vis_cfg.input_channels, vis_cfg.image_size, vis_cfg.image_size), device=device)
-                for _ in range(pad_length)
-            ]
-            padded_images.extend(padding)
-            presence_mask.extend([False] * pad_length)
-        image_batch.append(padded_images)
-        image_presence_mask.append(presence_mask)
-
-    # Convert lists to tensors
-    finalised_token_ids = torch.tensor(finalised_token_ids, dtype=torch.long, device=device)
-    if max_num_images > 0:
-        image_batch = torch.stack([torch.stack(images) for images in image_batch]).to(device)
-        image_presence_mask = torch.tensor(image_presence_mask, dtype=torch.bool, device=device)
-    else:
-        image_batch = None
-        image_presence_mask = None
-
-    # Prepare the output
-    output = TokenisationOutput(
-        finalised_token_ids=finalised_token_ids,
-        image_batch=image_batch,
-        batch_size=len(preprocessed_batch),
-        min_prompt_len=min_prompt_len,  #  type: ignore Intially set to float('inf'), but then we take a min
-        max_prompt_len=max_prompt_len,
-        max_seq_len=max_seq_len,
-        image_presence_mask=image_presence_mask,
-    )
-    return output
