+++ lxmls/multimodal/gemma3/siglip_vision/__init__.py
@@ -1,5 +1,4 @@
 from .config import SiglipVisionModelConfig
 from .model import SiglipVisionModel
-from .preprocessor import preprocess_images_for_siglip_vision
 
-__all__ = ["SiglipVisionModel", "SiglipVisionModelConfig", "preprocess_images_for_siglip_vision"]
+__all__ = ["SiglipVisionModel", "SiglipVisionModelConfig"]
